+++ src/main/java/models/interfaces/Vehicle.java
@@ -1,5 +1,6 @@
 package models.interfaces;
 
+import java.math.BigDecimal;
 import java.util.Date;
 
 public interface Vehicle {
@@ -7,11 +8,11 @@ public interface Vehicle {
     String getModel();
     Date getConstructionYear();
     String getColor();
-    double getPrice();
+    BigDecimal getPrice();
 
     void setManufacturer(Manufacturer manufacturer);
     void setModel(String model);
     void setConstructionYear(Date constructionYear);
     void setColor(String color);
-    void setPrice(double price);
+    void setPrice(BigDecimal price);
 }
